requirements.txt
argon2-cffi~=21.3.0
click==8.1.3
DateTime==4.4
dnspython==2.2.1
gunicorn==20.1.0
itsdangerous==2.1.2
jinja2==3.1.2
MarkupSafe==2.1.1
pymongo==4.1.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
Werkzeug==2.1.2
zope.interface==5.4.0
oauthlib~=3.2.0
pyOpenSSL==22.0.0
cryptography==37.0.2
argon2~=0.1.10
pip~=22.1.2
wheel~=0.37.1
lxml~=4.9.0
setuptools==62.6.0
requests~=2.25.1
google-oauth==1.0.0
google-auth==2.8.0
google==3.0.0
argparse~=1.4.0
mistune~=2.0.2
arrow~=1.2.2
starlette==0.20.3
typing-extensions==4.2.0
uvicorn~=0.18.1
aiohttp~=3.8.1
twilio~=7.9.3
websockets~=10.3